Refactor file names with PR suggestions in mind
