+++ week-1-ts-ecosystem/w01e01-typescript-basics/src/model/Address.ts
@@ -1,0 +1,6 @@
+interface Address {
+  street: string;
+  no: number;
+}
+
+export {Address}
